#23 combined some code for page and theme editors,
