+++ maqetta.core.client/WebContent/davinci/ve/commands/ModifyRuleCommand.js
@@ -9,9 +9,10 @@ return declare("davinci.ve.commands.ModifyRuleCommand", null, {
 
 	name: "modify rule",
 
-	constructor: function(cssRule, values){
+	constructor: function(cssRule, values, context){
 		this.cssRule = cssRule;
 		this.values = values;
+		this.context = context;
 		
 	},
 
@@ -27,7 +28,7 @@ return declare("davinci.ve.commands.ModifyRuleCommand", null, {
 		this.context.modifyRule( this.cssRule, this.values);
 		var file = this.cssRule.getCSSFile();
 		file.setDirty(true);
-		
+		this.context.editor.setDirty(true);
 		// Recompute styling properties in case we aren't in Normal state
 		States.resetState(this.context.rootNode);
 		if (this.context._selection) {
@@ -37,15 +38,14 @@ return declare("davinci.ve.commands.ModifyRuleCommand", null, {
 	},
 
 	undo: function(){
-		if(!this.cssRule || !this.values || !this.context){
+		if(!this.cssRule || !this._oldValues || !this.context){
 			return;
 		}
 
 		this.context.ruleSetAllProperties(this.cssRule, this._oldValues);
-		
 		var file = this.cssRule.getCSSFile();
 		file.setDirty(true);
-		
+		this.context.editor.setDirty(true);
 		// Recompute styling properties in case we aren't in Normal state
 		States.resetState(this.context.rootNode);
 		if (this.context._selection) {
