+++ maqetta.core.client/WebContent/davinci/ve/nls/common.js
@@ -4,6 +4,7 @@ define({ root:
 search: "Search",
 filter: "Filter",
 "new": "New",
+errorSavingFile: "Error saving file: ${0} details: ${1}",
 
 /* actions */
 undo: "Undo",
